LayoutTests/platform/glib/imported/w3c/web-platform-tests/webmessaging/with-ports/015-expected.txt

CONSOLE MESSAGE: Unable to post message to http://example.org. Recipient has origin http://web-platform.test:8800.


PASS different origin
